fix(hooks): run ESLint without a shell in the pre-commit hook
The hook built a `sh -c 'cd "<pkg>" && "<bin>" "<file>" …'` string per touched
package. lint-staged spawns that with no shell of its own, so the inner script
was interpreted by sh with the paths already embedded in it: any staged path
containing a double quote, `$` or a backtick broke the command, and a crafted
filename executed whatever it contained. Escaping through lint-staged's
string-argv parsing plus sh cannot be made watertight, so drop the shell.

The ESLint invocation moves to scripts/lint-staged-eslint.mjs, wired as a plain
string command so lint-staged appends the staged paths as separate argv entries.
The script groups them by package, then spawns the ESLint CLI through node with
`cwd` set — no path is ever interpolated into a command string. Behaviour is
otherwise unchanged: each package is still linted from its own directory so
ESLint 9 resolves that package's flat config, only staged files are passed, and
packages without a flat config are still skipped.

Verified: a staged file named `__probe";id>pwned-probe.txt;".ts` writes the file
under the old quoting and is linted as a literal path under the new runner; a
parse error still exits 1; multi-package and config-eslint-only runs behave as
before.
