+++ .lintstagedrc.mjs
@@ -1,48 +1,8 @@
-import { existsSync } from "node:fs";
-import path from "node:path";
-
-const ESLINT_BIN = "node_modules/.bin/eslint";
-
-// The `config-*` packages ship the shared presets themselves and have no
-// eslint.config.* of their own, so `turbo run lint` skips them. Running ESLint
-// there anyway aborts with "couldn't find an eslint.config file" and fails the
-// whole commit, so skip any package without a flat config.
-function hasEslintConfig(pkgDir) {
-  return ["mjs", "js", "cjs", "ts"].some((ext) => existsSync(path.join(pkgDir, `eslint.config.${ext}`)));
-}
-
-// Groups staged files by their owning app/package directory (e.g. "apps/web",
-// "packages/database"), each exactly one level under apps/ or packages/.
-function groupByPackage(files) {
-  const groups = new Map();
-  for (const absFile of files) {
-    const rel = path.relative(process.cwd(), absFile).split(path.sep).join("/");
-    const match = /^(apps|packages)\/([^/]+)\//.exec(rel);
-    if (!match) continue;
-    const pkgDir = `${match[1]}/${match[2]}`;
-    if (!groups.has(pkgDir)) groups.set(pkgDir, []);
-    groups.get(pkgDir).push(path.relative(pkgDir, rel));
-  }
-  return groups;
-}
-
-// Lints only the staged files themselves (not the whole package) by cd-ing into
-// each touched package and invoking its own eslint.config.mjs directly, so a
-// commit only pays for the files it touches and pre-existing violations
-// elsewhere in the package can't block it.
-function lintStagedFiles(files) {
-  const groups = groupByPackage(files);
-  return [...groups.entries()]
-    .filter(([pkgDir]) => hasEslintConfig(pkgDir))
-    .map(([pkgDir, relFiles]) => {
-      const eslintBin = path.relative(pkgDir, ESLINT_BIN);
-      const quotedFiles = relFiles.map((f) => `"${f}"`).join(" ");
-      return `sh -c 'cd "${pkgDir}" && "${eslintBin}" ${quotedFiles}'`;
-    });
-}
-
 export default {
-  "(apps|packages)/**/*.{js,ts,jsx,tsx,mjs}": ["prettier --write", lintStagedFiles],
+  // scripts/lint-staged-eslint.mjs receives the staged paths from lint-staged as
+  // argv, so nothing is interpolated into a shell command. It lints each touched
+  // package from that package's own directory.
+  "(apps|packages)/**/*.{js,ts,jsx,tsx,mjs}": ["prettier --write", "node scripts/lint-staged-eslint.mjs"],
   "*.json": ["prettier --write"],
   "packages/database/schema/**/*.prisma": ["prisma format"],
 };
